+++ src/main/java/bio/terra/workspace/app/StartupInitializer.java
@@ -1,6 +1,6 @@
 package bio.terra.workspace.app;
 
-import bio.terra.workspace.app.configuration.WorkspaceJdbcConfiguration;
+import bio.terra.workspace.app.configuration.JdbcConfiguration;
 import bio.terra.workspace.service.job.JobService;
 import bio.terra.workspace.service.migrate.MigrateService;
 import org.springframework.context.ApplicationContext;
@@ -11,14 +11,14 @@ public final class StartupInitializer {
   public static void initialize(ApplicationContext applicationContext) {
     // Initialize or upgrade the database depending on the configuration
     MigrateService migrateService = (MigrateService) applicationContext.getBean("migrateService");
-    WorkspaceJdbcConfiguration workspaceJdbcConfiguration =
-        (WorkspaceJdbcConfiguration) applicationContext.getBean("workspaceJdbcConfiguration");
+    JdbcConfiguration jdbcConfiguration =
+        (JdbcConfiguration) applicationContext.getBean("jdbcConfiguration");
     JobService jobService = (JobService) applicationContext.getBean("jobService");
 
-    if (workspaceJdbcConfiguration.isInitializeOnStart()) {
-      migrateService.initialize(changelogPath, workspaceJdbcConfiguration.getDataSource());
-    } else if (workspaceJdbcConfiguration.isUpgradeOnStart()) {
-      migrateService.upgrade(changelogPath, workspaceJdbcConfiguration.getDataSource());
+    if (jdbcConfiguration.isInitializeOnStart()) {
+      migrateService.initialize(changelogPath, jdbcConfiguration.getDataSource());
+    } else if (jdbcConfiguration.isUpgradeOnStart()) {
+      migrateService.upgrade(changelogPath, jdbcConfiguration.getDataSource());
     }
 
     // The JobService initialization also handles Stairway initialization.
