--- src/main/java/bio/terra/workspace/app/configuration/WorkspaceJdbcConfiguration.java
@@ -1,49 +1,0 @@
-package bio.terra.workspace.app.configuration;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.jdbc.datasource.DataSourceTransactionManager;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.PlatformTransactionManager;
-import org.springframework.transaction.annotation.EnableTransactionManagement;
-
-@Component
-@EnableConfigurationProperties
-@EnableTransactionManagement
-@ConfigurationProperties(prefix = "workspace.jdbc")
-public class WorkspaceJdbcConfiguration extends JdbcConfiguration {
-  // These properties control code in the StartupInitializer. We would not use these in production,
-  // but they
-  // are handy to set for development and testing. There are only three interesting states:
-  // 1. initialize is true; upgrade is irrelevant - initialize and recreate an empty database
-  // 2. initialize is false; upgrade is true - apply changesets to an existing database
-  // 3. initialize is false; upgrade is false - do nothing to the database
-  /** If true, primary database will be wiped */
-  private boolean initializeOnStart;
-  /** If true, primary database will have changesets applied */
-  private boolean upgradeOnStart;
-
-  public boolean isInitializeOnStart() {
-    return initializeOnStart;
-  }
-
-  public void setInitializeOnStart(boolean initializeOnStart) {
-    this.initializeOnStart = initializeOnStart;
-  }
-
-  public boolean isUpgradeOnStart() {
-    return upgradeOnStart;
-  }
-
-  public void setUpgradeOnStart(boolean upgradeOnStart) {
-    this.upgradeOnStart = upgradeOnStart;
-  }
-
-  // This bean plus the @EnableTransactionManagement annotation above enables the use of the
-  // @Transaction annotation to control the transaction properties of the data source.
-  @Bean("transactionManager")
-  public PlatformTransactionManager getTransactionManager() {
-    return new DataSourceTransactionManager(getDataSource());
-  }
-}
