+++ src/test/resources/config/application-integration-test.yml
@@ -7,6 +7,6 @@ workspace.integration-test:
   service-account-file-path: rendered/service-account.json
   wsm-endpoints:
     workspaces: api/workspaces/v1
+  wsm-urls:
     dev: https://workspace.dsde-dev.broadinstitute.org/
 #    dev: http://localhost:8080/
-  wsm-urls:
